[BEAM-6683] add createCrossLanguageValidatesRunner task

+++ sdks/python/apache_beam/transforms/external_test.py
@@ -25,22 +25,24 @@ import sys
 import unittest
 
 import grpc
+from nose.plugins.attrib import attr
 from past.builtins import unicode
 
 import apache_beam as beam
 from apache_beam import Pipeline
-from apache_beam.io.external.generate_sequence import GenerateSequence
-from apache_beam.portability import python_urns
 from apache_beam.runners.portability import expansion_service
 from apache_beam.runners.portability.expansion_service_test import FibTransform
+from apache_beam.testing.test_pipeline import TestPipeline
 from apache_beam.testing.util import assert_that
 from apache_beam.testing.util import equal_to
 
 
+@attr('UsesCrossLanguageTransforms')
 class ExternalTransformTest(unittest.TestCase):
 
   # This will be overwritten if set via a flag.
   expansion_service_jar = None
+  expansion_port = None
 
   def test_pipeline_generation(self):
     pipeline = beam.Pipeline()
@@ -103,33 +105,31 @@ class ExternalTransformTest(unittest.TestCase):
       assert_that(p | FibTransform(6), equal_to([8]))
 
   def test_java_expansion(self):
-    if not self.expansion_service_jar:
-      raise unittest.SkipTest('No expansion service jar provided.')
+    test_pipeline = TestPipeline()
+
+    self.expansion_port = test_pipeline.get_option("expansion_port")
+
+    if not (self.expansion_service_jar or self.expansion_port):
+      raise unittest.SkipTest('No expansion service jar or port provided.')
 
     # The actual definitions of these transforms is in
     # org.apache.beam.runners.core.construction.TestExpansionService.
-    TEST_COUNT_URN = "pytest:beam:transforms:count"
-    TEST_FILTER_URN = "pytest:beam:transforms:filter_less_than"
+    TEST_COUNT_URN = "beam:transforms:xlang:count"
+    TEST_FILTER_URN = "beam:transforms:xlang:filter_less_than_eq"
 
-    # Run as cheaply as possible on the portable runner.
-    # TODO(robertwb): Support this directly in the direct runner.
-    options = beam.options.pipeline_options.PipelineOptions(
-        runner='PortableRunner',
-        experiments=['beam_fn_api'],
-        environment_type=python_urns.EMBEDDED_PYTHON,
-        job_endpoint='embed')
-
+    server = None
     try:
-      # Start the java server and wait for it to be ready.
-      port = '8091'
+      port = self.expansion_port or '8091'
       address = 'localhost:%s' % port
-      server = subprocess.Popen(
-          ['java', '-jar', self.expansion_service_jar, port])
+      if self.expansion_service_jar:
+        # Start the java server and wait for it to be ready.
+        server = subprocess.Popen(
+            ['java', '-jar', self.expansion_service_jar, port])
       with grpc.insecure_channel(address) as channel:
         grpc.channel_ready_future(channel).result()
 
       # Run a simple count-filtered-letters pipeline.
-      with beam.Pipeline(options=options) as p:
+      with test_pipeline as p:
         res = (
             p
             | beam.Create(list('aaabccxyyzzz'))
@@ -144,18 +144,9 @@ class ExternalTransformTest(unittest.TestCase):
 
         assert_that(res, equal_to(['a: 3', 'b: 1', 'c: 2']))
 
-      # Test GenerateSequence Java transform
-      with beam.Pipeline(options=options) as p:
-        res = (
-            p
-            | GenerateSequence(start=1, stop=10,
-                               expansion_service=address)
-        )
-
-        assert_that(res, equal_to([i for i in range(1, 10)]))
-
     finally:
-      server.kill()
+      if server:
+        server.kill()
 
 
 if __name__ == '__main__':
